backend/requirements.txt
asgiref
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse
tzdata
django-storages
boto3
dj-database-url==2.3.0
gunicorn
whitenoise
PyJWT
Pillow
supabase
python-dotenv
django-debug-toolbar
stripe
requests
django-jazzmin==2.6.0